pyproject.toml
[project]
name = "zotero-mcp"
version = "0.1.1"
description = "Model Context Protocol server for Zotero"
authors = [{ name = "Aaron Taylor", email = "pypi@ataylor.io" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = ["mcp", "zotero"]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp[cli]>=1.2.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pyzotero>=1.6.8",
]

[project.scripts]
zotero-mcp = "zotero_mcp.cli:main"

[project.urls]
Repository = "https://github.com/kujenga/zotero-mcp"
Issues = "https://github.com/kujenga/zotero-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
